Use c.BuildContext in matchPackagesInFS
In the original cmd/go code, a package scoped unexported variable
buildContext is used. In the gotool version, that variable is replaced
by c.BuildContext field in most places, except this one.

The change in #8 partially introduced this fix, but had to be reverted
in #11 due to build errors.

This reintroduces the fix while resolving build errors.