Add gotool implementation for Go 1.9.
Go 1.9 has modified cmd/go by splitting it into multiple internal
packages. The functionality we care about is largely contained in
the cmd/go/internal/load package. We make a copy of that package
and use it for Go 1.9.

The ./internal/load package is a copy of cmd/go/internal/load
package. It is modified in a minimal way. Specifically:

-	There are +build go1.9 constraints added.
-	Context struct is added, and functions that need to
	access its fields are made methods of it.
-	Unused code in pkg.go and path.go files is not included.

The configuration values (build context, location of GOROOT/src
directory) that are normally accessed from cmd/go/internal/cfg
package are instead provided via load.Context.

The goal of this change was to keep diff of ./internal/load package
against cmd/go/internal/load minimal, to make it easiest to keep it
in sync and to have high confidence in its behavior being equivalent.